RDKUI-852: Correctly set language during FTI
Previously, when index was being updated when user scrolled to "Continue".
It caused to save Spanish language as selected, as it was most recently selected
langauge.

Currently, an aditional variable is used as a localstorage that is updated on "enter".
It is used to indicate what language was selected. If this is undefined, use 0 index which
is the default. Localstorage is used to persist after page reload.